requirements.txt
polars==1.27.0
psycopg2==2.9.10
py4j==0.10.9.7
python-dotenv==1.1.0
SQLAlchemy==2.0.40
tqdm==4.67.1
typing_extensions==4.13.1